bot/requirements.txt
apscheduler==3.6.3
cachetools==4.2.2
certifi==2021.10.8
peewee==3.14.8
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2021.5
tzlocal==4.1
